Day 19 part 2

+++ src/main/kotlin/com/rolf/day19/Part.kt
@@ -1,12 +1,12 @@
 package com.rolf.day19
 
 data class Part(
-    val x: Long,
-    val m: Long,
-    val a: Long,
-    val s: Long,
+    val x: Int,
+    val m: Int,
+    val a: Int,
+    val s: Int,
 ) {
-    fun sum(): Long {
+    fun sum(): Int {
         return x + m + a + s
     }
 }
